requirements.txt
allure-pytest==2.13.2
allure-python-commons==2.13.2
asttokens==2.2.1
async-generator==1.10
attrs==22.2.0
backcall==0.2.0
certifi==2022.12.7
charset-normalizer==3.1.0
comm==0.1.3
debugpy==1.6.7
decorator==5.1.1
exceptiongroup==1.1.1
execnet==1.9.0
executing==1.2.0
h11==0.14.0
idna==3.4
iniconfig==2.0.0
ipykernel==6.22.0
ipython==8.12.0
jedi==0.18.2
jupyter_client==8.2.0
jupyter_core==5.3.0
Mako==1.2.4
MarkupSafe==2.1.3
matplotlib-inline==0.1.6
mysql-connector-python==8.0.33
nest-asyncio==1.5.6
outcome==1.2.0
packaging==23.1
parse==1.19.0
parse-type==0.6.0
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
platformdirs==3.2.0
pluggy==1.0.0
prompt-toolkit==3.0.38
protobuf==3.20.3
psutil==5.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
Pygments==2.15.0
PySocks==1.7.1
pytest==7.3.1
pytest-bdd==6.1.1
pytest-rerunfailures==11.1.2
pytest-xdist==3.3.1
python-dateutil==2.8.2
pyzmq==25.0.2
requests==2.31.0
selenium==4.9.1
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
stack-data==0.6.2
tabulate==0.9.0
tomli==2.0.1
tornado==6.3
traitlets==5.9.0
trio==0.22.0
trio-websocket==0.10.2
typing_extensions==4.6.3
urllib3==1.26.15
wcwidth==0.2.6
wsproto==1.2.0
